couple minor tweaks
